Rename storage module
