+++ bee-storage/bee-storage/src/access/fetch.rs
@@ -1,7 +1,7 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::storage::StorageBackend;
+use crate::backend::StorageBackend;
 
 /// Fetch<K, V> trait will extend the StorageBackend with Fetch operation for the key: K value: V pair
 /// therefore it should be explicitly implemented for the corresponding StorageBackend.
