+++ bee-storage/bee-storage/src/access/truncate.rs
@@ -1,7 +1,7 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::storage::StorageBackend;
+use crate::backend::StorageBackend;
 
 #[async_trait::async_trait]
 pub trait Truncate<K, V>: StorageBackend {
